Add BashTest
